Security for all page

+++ src/main/java/control/DeleteAccountControl.java
@@ -8,6 +8,7 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 
 import dao.DAO;
 import entity.Account;
@@ -21,12 +22,29 @@ public class DeleteAccountControl extends HttpServlet {
 	protected void processRequest(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
 		
-		String aid = request.getParameter("aid");
-		System.out.println(aid);
+		HttpSession session = request.getSession();
+        Account user = (Account) session.getAttribute("acc"); 
         DAO dao = new DAO();
-        dao.deleteAccount(aid);
         
-     	response.sendRedirect("managerAccount"); // Parameter here is the URL Pattern: /manager 
+        if (user != null) {
+            if (user.getIsAdmin() == 1 ) {
+            	String aid = request.getParameter("aid");
+        		System.out.println(aid);
+                dao.deleteAccount(aid);
+                
+             	response.sendRedirect("managerAccount"); 
+            } else {
+            	request.getRequestDispatcher("/WEB-INF/views/AccessDenied.jsp").forward(request, response); // Chuyển hướng đến trang từ chối truy cập
+            }
+        } else {
+            response.sendRedirect("login"); // Chuyển hướng đến trang đăng nhập
+        }
+//		String aid = request.getParameter("aid");
+//		System.out.println(aid);
+//        DAO dao = new DAO();
+//        dao.deleteAccount(aid);
+//        
+//     	response.sendRedirect("managerAccount"); // Parameter here is the URL Pattern: /manager 
     }
 
     @Override
